+++ sbnz-integracija-projekta/model/src/main/java/com/ftn/sbnz/model/models/FaultProblems/MotorProblem.java
@@ -5,11 +5,11 @@ public class MotorProblem extends PowertrainProblems{
     private int responsivness;
 
 
-    public MotorProblem(long breakdownId) {
+    public MotorProblem(int breakdownId) {
         super(breakdownId);
     }
 
-    public MotorProblem(int responsivness, long breakdownId) {
+    public MotorProblem(int responsivness, int breakdownId) {
         this(breakdownId);
         this.responsivness = responsivness;
     }
