+++ sbnz-integracija-projekta/model/src/main/java/com/ftn/sbnz/model/models/FaultProblems/WheelSpeedProblem.java
@@ -4,11 +4,11 @@ import java.util.Objects;
 public class WheelSpeedProblem extends RegenerativeBreakingProblem{
     private double matching;
 
-    public WheelSpeedProblem(long breakdownId) {
+    public WheelSpeedProblem(int breakdownId) {
         super(breakdownId);
     }
 
-    public WheelSpeedProblem(double matching, long breakdownId) {
+    public WheelSpeedProblem(double matching, int breakdownId) {
         this(breakdownId);
         this.matching = matching;
     }
